requirements.txt
absl-py
jax
jaxlib
dm-env
dm-haiku
bsuite
 